+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -3,6 +3,7 @@
 - Improvement: Added `AddNestedRecord` to `CSharpClass`.
 - Improvement: Clears InteractionStrategy cache between "Hot" Software Factory restarts.
 - Improvement: Added `Blank line between Members` setting to allow for "always" adding a blank line between members.
+- Improvement: Elements in the `Domain` and `Services` designers will be scanned for the presence of `C#` stereotypes and add known C# types for purposes of disambiguation.
 
 ### Version 3.9.5